M0.0.1S - Initial Commit
This is the initial commit of the 1923 2022 offseason build. I started to follow the "Zero to Autonomous" swerve tutorial and have made some adjustments since we are using Falcons instead of the Neos that they use in the video. Notable changes include the UnitConversion class which contains ticksToRotations, rotationsToTicks, and ticksToRadians methods.