B0.0.2S - Zero to Autonomous Cloned
Before testing on 5/24/22, I finished porting over the code from the Zero to Autonomous tutorial to something that can hopefully be useable. It follows almost exactly as is seen in their video. Anything that I was unsure of, especially in Constants, was denoted with a "?" in the comment next to it. These values should be assessed and updated as required.