B0.1.0SM - Swerve Logic Functional
The swerve drive code has been tested and seems to be almost perfectly working. State generation for each module seems to be correct as well as motor drive speeds, but currently the motors are rotating to the wrong positions. The turning PID is just going to the wrong setpoints. The setpoints are correct, but I feel that the logic to get the goal the motor needs to rotate to is inverted somehow. Maybe only needs one/two things to be changed to have a working swerve.

Current button layout is as follows:

- Left stick is translation (X and Y axis)
- Right stick is rotation (X axis)
- Right Bumper while held enables robot-centric driving (normally field centric)