B0.1.1SM - Possible Fix for Swerve Issues
I think I might have fixed our swerve offset issue. I felt that the offsets were never actually being used so I did a dive in the code and I was right. The only time we access the sensor is to print it out to SmartDashboard! So, I found where we reset our encoders and I discovered we are just resetting the turning motor to be equal to an amount of ticks equivlent to the offset radians we set in Constants. That is wrong as it would not matter where the wheel actually was, and it would just reset from that point but skewed. I rewatched the Zero to Autonomous tutorial and I was right, they use the getter method for their absolute encoder in the resetting encoder method. Whoops! Still needs to be tested.