B0.2.0SM - Teleop Swerve Drive Functional
The resetting issue was fixed with the last few pushes, however I realized that the Falcon's encoder ticks are the opposite direction of the absolute encoders, so I set them all to be inverted after we remeasured the absolute encoder offsets. Also, I had to invert the drive direction on all motors to get it to drive the correct direction. Finally, I had to invert the gyro heading  because the field-oriented code was accessing it opposite of what it should be (turning the robot made the wheels turn even more in that direction, not opposite to cancel out the rotation).

Afterwards, we had to swap the button ID for the robot centric mode because of the fact that button ID's start at 1, not 0.

Also swapped the team number in wpilib_preferences to 1923.

Finally, I added LimelightInterface and LimelightSubsystem classes from old 747 code to make accessing it easier. I created the GoalCentricCommand which allows for field-centric translation with rotation being controlled by a kP times the feedback from the vision tracking camera, keeping the robot always facing the goal. We need to make the wheels not so snappy in that mode, as driving is affected, as well as we want to increase the kP from .01 -> .02 to make it a little more snappy.

Right bumper is robot centric while in the default command, and the "A" button enables GoalCentricCommand.

The gyro is reset on boot/reset robot code and it must be set up in a way that the robot is facing where you want away to be when you push the stick forward. Otherwise the robots field of view is off. Spinning for long periods of time without resetting gyro accumulates gyro drift and the robot might drive offset slightly as time goes on.