B0.3.0SM - Trajectory Following Started
I added the required constants and methods to the swerve subsystem and constants class to complete the swerve video from zero to autonomous. Currently, the robot will follow the trajectory they suggested in the video decently. The robot does not go the full distance required and it does not turn the full angle that is requested, but it is definitely making an attempt to go to the right heading.

Halfway through testing driving seemed to become more jittery (as if a wheel was not aligned properly). Examination of the wheels seemed to show that all wheels were facing the correct direction. Eventually, the bumping got significantly worse and then the tread of the front left wheel ripped itself off. So far, we are believing that the tread issue caused the wonky driving, not that the wonky driving caused the tread issue.

I leave off today wondering where robot characterization would play a role in the autonomous of the robot, as well as trying to find a way to generate on-the-fly trajectories so we can use it at competition.