B0.3.1SM - PathPlanner Added
I took some time to convert the FollowTrajectory command group to be usable for any path you want. Using the PathPlanner library and code from  @MC42 and @Squid5678, I created a simple implementation of a class that can handle "any" path.

Auto still needs to be tuned and none of this was actually tested (I have some doubts about if the robot will be able to find the path based on how I typed it), but overall this isn't a huge deviation from what I saw working in person today. The path the robot will drive is a 2 meter drive straight forward. I assume we have to tune the 3 PID controllers independently, so I think three different paths for tuning is appropriate. Thought I would write this at home so we didn't waste time during the next meeting.