B0.3.2SM - XController tuned, trouble with Y
I tuned the X PID controller by setting Y and theta controllers to zero, then giving it the 2 meter path to follow. We tuned that and the wheel diameter to make it go the correct distance (had to lower diameter to 3.7" from 4").

Then, I moved to the Y controller. I set X and theta to 0 (but I saved in a comment the X value) and told the robot to strafe 2 meters to the left while maintaining a rotation of looking forward. The robot instead did the same physical path as the X path (2 meters forward). The odometry however showed an increase in movement in the X direction. This seems to be an issue with how heading works in the path planner program? I had to make the heading face up, instead of outwards, since otherwise the spline that was generated would not be straight. The confusing thing is that path planner has both "heading" and "rotation" arguments, as well as a heading magnitude which you have to do by hand and does not have an input.

Three potential ideas for fixes:

1. Implement the PathPlanner swerve auto command and see if there was just a different way they handle trajectories over the WPILib version

2. Create a trajectory object that should strafe left or right in code (WPILib trajectory object)

3. Manually edit the splines to face the same way as the robot heading using the number inputs, but drag the magnitude so the spline has no curvature (has to be done by eye)

Hopefully one of these will fix the issue. Will write code later and push for Friday testing.