B0.4.0SM - Modules on Real Frame & MKI Planner Classes
All of the modules have been mounted & wired on the new frame and the code has been adjusted to support them.

In terms of autonomous, we spent a lot more time testing to figure out how to achieve the desired strafing effect. We are still unsure if this is an issue with the wheels turning with PID or a trajectory generation loophole. To fix the fact that the PathPlanner trajectories were mirrored in the Y-direction, I copied their classes that we used and made MKI versions where the Y values are negative when passed in, fixing the issue.