B0.4.5SM - Cleaning & New Hood Position Command
Created a new command akin to 747 PIDDriveInches command written by @BNJimenez . It takes in a position goal and the command will exit once you are within a threshold for a certain amount of time. Should work similar to the DART in 2019 from 747. Otherwise, did a lot of cleaning and organization to improve workflow within the other subsystems. Also believe the motor issue we were having was caused by the fact that multiple motors in constants were assigned to ID 0. Finally, added a few new helper methods and SmartDashboard printouts for the next meeting,