B0.4.7SM - Created Hood Subsystem
I created a HoodSubsystem and separated all necessary code from the ShooterSubsystem.

I did this because I think that during our control scheme we are going to want a default command for both the shooter and the hood that are going to be interrupted and replaced at different times. There is probably a way to do that with a single subsystem, but I figured just using the standard subsystem framework was the best option.