B0.4.8SM - Refactoring & Shooter Command
Moved around some commands to organize a little better. Also, created a RunShooterCommand that will maintain a goal RPM for the shooter. It will not exit. It also will manage a boolean in ShooterSubsystem that we can access in our other code. It represents if the shooter has been within an acceptable range of RPM for a period of time we define.