B0.4.9SM -  Shooter & Hood Commands
Wrote some more complicated commands to handle the actual motion we expect to see during teleop. Also created an avoid stall default command for the shooter, where it will maintain a speed throughout the match in order to reduce current draw when shooting.

Added a map-onto-range function and angle-to-target function to the LimelightSubsystem.

Did some name refactoring as well. Renamed stop functions to just be stop() in the respective subsystems to avoid confusion. Also, changed some singular nouns to plural where required.