B0.5.0SM - StateHandler
Wrote the StateHandler class and haven't had any time to test it. This should be a fine way to easily measure the current ball state on the robot, for future use in our intake shoot commands. Currently, the logic is untested, but should be a good starting point. Finally, I added the method updateBooleans to a 20ms loop in Robot.java. This is used to update the local booleans in the StateHandler to avoid the case where booleans are partially updated when it is calculating the new state, providing an incorrect state calculation.