B0.6.2SM - Gyro Odometry Filtering?
Wrote an extension to the new MKIPoseEstimator. This will take in the gyro velocities (x and y assuming those are the relevant 2d components) and use how much it differs from the encoder values to offset our outputs. This will hopefully improve our pose estimation. However, for this to work effectively, we need to change the conversion ratio between the native gyro velocity to meters-per-second.