B0.6.4.1AT - Shooter Tuning and Updated Climb Sequence for Level 3
We started today off working with the climb sequence. The hood now moves in parallel when climbing level 2, making it significantly faster. We were able to achieve 1 level 3 climb today before the tensioning on the climber screwed up and we were no longer able to test our climb. We then went to test our shooter accuracy with the limelight. It's still overshooting a bit and some more turning is needed, but it seems to work fine at the moment.