B0.7.0SM - Rem. Garbage & MKITrajectoryGenerator
Added our logic to easily define a starting & ending velocity for trajectories. Based on advice from 341 (Miss Daisy). Had to create a MKITrajectoryConfig class to allow for this to work. Just a dependency.

Also, removed the path planner trajectories and made a new folder to store new trajectories called "routines" in deploy.