B1.5AT - Coach Dashboard
Added a coach dashboard onto shuffleboard. I left the auto chooser out on the other "driver dashboard" page for now since it's not needed, we can add it later. We also need to add the Limelight but is there any way we can do it without the manual drop-in? (currently WPILib documentation says no).